Separate Monolithic `Init.lua` file to multiple files. (#1)
## Commits made:
* Create necessary files and folders, and also...

...added the require(file) statements in the init.lua file.

* Fill in keymaps.lua and settings.lua

These were done together because they are much less complicated to do.

* Add plugins.lua and moved the install-lazy autocmd

* Attempt to fix Plugin require-ing (WIP)

* Fix minor typo

* Fix up organization-attempt by referencing repo

I am referencing the modular kickstart.nvim repo.

* Rename plugins to lazy-plugins & remove lazy bootstrap

* Remove .lua extension from require function

I included those thinking that those were simply optional and ommitted
in the repo.
They were in fact, not optional and ommitted because it doesn't work
otherwise.

* Remove the trailing comma and table from the end.

This was causing an error.

* Fix nerd font being turned off accidentally.

* Disable auto-reloading plugins

I have autosave turned on, so this gets way too annoying too fast.

* Separate plugins/init.lua file into multiple files

The separation into multiple files was only done for plugins that
required more than a single line to import.

* Delete the monolithic init.lua file.

It is no longer needed. Goodbye.